+++ plugin/quartz/gvdevice_quartz.c
@@ -42,6 +42,17 @@ static void quartz_format(GVJ_t *job)
 	    format_to_uti((format_type)job->device.id), 1, NULL);
 	CGDataProviderRef data_provider = CGDataProviderCreateDirect(job->imagedata, BYTES_PER_PIXEL * job->width * job->height, &memory_data_provider_callbacks);
 	
+	CFStringRef keys[] = {kCGImagePropertyDPIWidth, kCGImagePropertyDPIHeight};
+	CFNumberRef values[] = {
+		CFNumberCreate(NULL, kCFNumberDoubleType, &job->dpi.x),
+		CFNumberCreate(NULL, kCFNumberDoubleType, &job->dpi.y)
+	};
+	CFDictionaryRef dpi = CFDictionaryCreate(NULL, (const void **)keys,
+	                                         (const void **)values,
+	                                         sizeof(keys) / sizeof(keys[0]),
+	                                         &kCFTypeDictionaryKeyCallBacks,
+	                                         &kCFTypeDictionaryValueCallBacks);
+
 	/* add the bitmap image to the destination and save it */
 	CGColorSpaceRef color_space = CGColorSpaceCreateWithName(kCGColorSpaceSRGB);
 	CGImageRef image = CGImageCreate (
@@ -57,7 +68,7 @@ static void quartz_format(GVJ_t *job)
 		false, // don't interpolate
 		kCGRenderingIntentDefault			/* rendering intent (what to do with out-of-gamut colors): default */
 	);
-	CGImageDestinationAddImage(image_destination, image, NULL);
+	CGImageDestinationAddImage(image_destination, image, dpi);
 	CGImageDestinationFinalize(image_destination);
 	
 	/* clean up */
@@ -66,6 +77,8 @@ static void quartz_format(GVJ_t *job)
 	CGDataProviderRelease(data_provider);
 	if (image_destination)
 		CFRelease(image_destination);
+	if (dpi != NULL)
+		CFRelease(dpi);
 	CGDataConsumerRelease(data_consumer);
 }