+++ plugin/quartz/gvloadimage_quartz.c
@@ -93,7 +93,7 @@ static CGImageRef quartz_loadimage(GVJ_t * job, usershape_t *us)
 		}
 		CFStringRef uti_hint = format_to_uti(hint_format_type);
 		CFDictionaryRef options = hint_format_type == FORMAT_NONE ? NULL : CFDictionaryCreate(
-			kCFAllocatorDefault,
+			NULL,
 			(const void **)&kCGImageSourceTypeIdentifierHint,
 			(const void **)&uti_hint,
 			1,